improv(frontend): remove search journey from modal
